Authentication working with Budget now To Be Completed: - Test make sure nothing has broken - Logout - Log in - Client side
--------
Have not dev'd on this in a while, which is why I am leaving the above. The user endpoint is working which can be seen in the endPointCalls.js file, appending headers.

I need to create a .env file and store all the database stuff in there, removing the test_database.js file as I believe this is where it is failing. Adding the Ticket to the board for visibility